backend/requirements.txt
# Backend Dependencies
# Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Google Services
google-generativeai>=0.4.0
gspread>=6.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0

# Geospatial
geopy>=2.4.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# HTTP client
httpx>=0.26.0